+++ src/traces/histogram/attributes.js
@@ -9,7 +9,9 @@
 'use strict';
 
 var barAttrs = require('../bar/attributes');
+var hovertemplateAttrs = require('../../components/fx/hovertemplate_attributes');
 var makeBinAttrs = require('./bin_attributes');
+var constants = require('./constants');
 
 module.exports = {
     x: {
@@ -184,6 +186,10 @@ module.exports = {
             'before deleting `autobiny` from the trace.'
         ].join(' ')
     },
+
+    hovertemplate: hovertemplateAttrs({}, {
+        keys: constants.eventDataKeys
+    }),
 
     marker: barAttrs.marker,