+++ test/jasmine/tests/bar_test.js
@@ -17,9 +17,14 @@ var supplyAllDefaults = require('../assets/supply_defaults');
 var color = require('../../../src/components/color');
 var rgb = color.rgb;
 
+var checkEventData = require('../assets/check_event_data');
+var constants = require('@src/traces/bar/constants');
+
 var customAssertions = require('../assets/custom_assertions');
 var assertClip = customAssertions.assertClip;
 var assertNodeDisplay = customAssertions.assertNodeDisplay;
+var assertHoverLabelContent = customAssertions.assertHoverLabelContent;
+var Fx = require('@src/components/fx');
 
 var d3 = require('d3');
 
@@ -1083,6 +1088,7 @@ describe('A bar plot', function() {
         .catch(failTest)
         .then(done);
     });
+
     it('should show bar texts (outside case)', function(done) {
         var data = [{
             y: [10, -20, 30],
@@ -1736,6 +1742,39 @@ describe('A bar plot', function() {
         .catch(failTest)
         .then(done);
     });
+
+    it('should be able to react with new text colors', function(done) {
+        Plotly.react(gd, [{
+            type: 'bar',
+            y: [1, 2, 3],
+            text: ['A', 'B', 'C'],
+            textposition: 'inside'
+        }])
+        .then(assertTextFontColors(['rgb(255, 255, 255)', 'rgb(255, 255, 255)', 'rgb(255, 255, 255)']))
+        .then(function() {
+            gd.data[0].insidetextfont = {color: 'red'};
+            return Plotly.react(gd, gd.data);
+        })
+        .then(assertTextFontColors(['rgb(255, 0, 0)', 'rgb(255, 0, 0)', 'rgb(255, 0, 0)']))
+        .then(function() {
+            delete gd.data[0].insidetextfont.color;
+            gd.data[0].textfont = {color: 'blue'};
+            return Plotly.react(gd, gd.data);
+        })
+        .then(assertTextFontColors(['rgb(0, 0, 255)', 'rgb(0, 0, 255)', 'rgb(0, 0, 255)']))
+        .then(function() {
+            gd.data[0].textposition = 'outside';
+            return Plotly.react(gd, gd.data);
+        })
+        .then(assertTextFontColors(['rgb(0, 0, 255)', 'rgb(0, 0, 255)', 'rgb(0, 0, 255)']))
+        .then(function() {
+            gd.data[0].outsidetextfont = {color: 'red'};
+            return Plotly.react(gd, gd.data);
+        })
+        .then(assertTextFontColors(['rgb(255, 0, 0)', 'rgb(255, 0, 0)', 'rgb(255, 0, 0)']))
+        .catch(failTest)
+        .then(done);
+    });
 });
 
 describe('bar visibility toggling:', function() {
@@ -2007,6 +2046,34 @@ describe('bar hover', function() {
             .catch(failTest)
             .then(done);
         });
+
+        it('should use hovertemplate if specified', function(done) {
+            gd = createGraphDiv();
+
+            var mock = Lib.extendDeep({}, require('@mocks/text_chart_arrays'));
+            mock.data.forEach(function(t) {
+                t.type = 'bar';
+                t.hovertemplate = '%{y}<extra></extra>';
+            });
+
+            function _hover() {
+                var evt = { xpx: 125, ypx: 150 };
+                Fx.hover('graph', evt, 'xy');
+            }
+
+            Plotly.plot(gd, mock)
+            .then(_hover)
+            .then(function() {
+                assertHoverLabelContent({
+                    nums: ['1', '2', '1.5'],
+                    name: ['', '', ''],
+                    axis: '0'
+                });
+                // return Plotly.restyle(gd, 'text', ['APPLE', 'BANANA', 'ORANGE']);
+            })
+            .catch(failTest)
+            .then(done);
+        });
     });
 
     describe('with special width/offset combinations', function() {
@@ -2230,6 +2297,11 @@ describe('bar hover', function() {
         .catch(failTest)
         .then(done);
     });
+});
+
+describe('event data', function() {
+    var mock = require('@mocks/stacked_bar');
+    checkEventData(mock, 216, 309, constants.eventDataKeys);
 });
 
 function mockBarPlot(dataWithoutTraceType, layout) {