+++ test/jasmine/tests/gl3d_plot_interact_test.js
@@ -558,6 +558,61 @@ describe('Test gl3d plots', function() {
         .catch(failTest)
         .then(done);
     });
+
+    it('@gl should only accept texts for textposition otherwise textposition is set to middle center before passing to webgl', function(done) {
+        Plotly.plot(gd, [{
+            type: 'scatter3d',
+            mode: 'markers+text+lines',
+            x: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+            y: [-1, -2, -3, -4, -5, -6, -7, -8, -9, -10, -11, -12, -13, -14, -15, -16],
+            z: [0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1],
+            text: ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p'],
+            textposition: ['left top', 'right top', 'left bottom', 'right bottom', null, undefined, true, false, [], {}, NaN, Infinity, 0, 1.2]
+        }])
+        .then(function() {
+            var AllTextpositions = gd._fullData[0].textposition;
+
+            expect(AllTextpositions[0]).toBe('top left', 'is not top left');
+            expect(AllTextpositions[1]).toBe('top right', 'is not top right');
+            expect(AllTextpositions[2]).toBe('bottom left', 'is not bottom left');
+            expect(AllTextpositions[3]).toBe('bottom right', 'is not bottom right');
+            for(var i = 4; i < AllTextpositions.length; i++) {
+                expect(AllTextpositions[i]).toBe('middle center', 'is not middle center');
+            }
+        })
+        .catch(failTest)
+        .then(done);
+    });
+
+    it('@gl axis ticks should not be set when axis _length is NaN', function(done) {
+        Plotly.plot(gd,
+            {
+                data: [{
+                    type: 'scatter3d',
+                    mode: 'markers',
+                    x: [1, 2],
+                    y: [3, 4],
+                    z: [5, 6]
+                }],
+                layout: {
+                    scene: {
+                        camera: {
+                            eye: {x: 1, y: 1, z: 0},
+                            center: {x: 0.5, y: 0.5, z: 1},
+                            up: {x: 0, y: 0, z: 1}
+                        }
+                    }
+                }
+            }
+        )
+        .then(function() {
+            var zaxis = gd._fullLayout.scene.zaxis;
+            expect(isNaN(zaxis._length)).toBe(true);
+            expect(zaxis.dtick === undefined).toBe(true);
+        })
+        .catch(failTest)
+        .then(done);
+    });
 });
 
 describe('Test gl3d modebar handlers', function() {